MAINTAINERS: Use Linux Kernel formatting style
The current markdown version of the file is not formatted properly
in the GitHub web viewer. It also lacks the scopes (files) for each
maintainer to take care of.

As such, remove the markdown file, replace it with the plaintext one,
and add in some properties (maintainer status, files) like the kernel.

Signed-off-by: Igor Belwon <igor.belwon@mentallysanemainliners.org>